Created Author Repository
All Author related actions on the database now happen through an IAuthorRepository instead of a ICheepRepository
