+++ src/Chirp.Infrastructure/Chirp.Repositories/IAuthorRepository.cs
@@ -1,0 +1,9 @@
+namespace Chirp.Repositories;
+using Core;
+
+public interface IAuthorRepository
+{
+    Task<AuthorDTO?> GetAuthorByName(string authorName);
+    Task<AuthorDTO?> GetAuthorByEmail(string authorEmail);
+    Task CreateAuthor(string authorName,string authorEmail);
+}
