+++ src/Chirp.Infrastructure/Chirp.Repositories/ICheepRepository.cs
@@ -6,9 +6,5 @@ public interface ICheepRepository
     Task CreateCheep(CheepDTO cheep);
     Task<List<CheepDTO>> ReadCheeps(string? user, int offset, int count);
     Task UpdateCheep(CheepDTO cheep);
-    Task<AuthorDTO?> GetAuthorByName(string authorName);
-    Task<AuthorDTO?> GetAuthorByEmail(string authorEmail);
-    Task CreateAuthor(string authorName,string authorEmail);
-    
     
 }
